requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
openai==0.27.0
requests==2.31.0
ollama
groq
transformers
torch